# Faster cofactor clearing for G1 & G2 of bls12-381 + benchmarking