## Description

Benchmarking results: 
G1: 2x speedup
- (h_eff cofactor-clearing) `test bls12_381::g1::cofactor_clearing  ... bench:      48,520 ns/iter (+/- 660)`
- (naive, mul by cofactor) `test bls12_381::g1::cofactor_clearing  ... bench:     111,015 ns/iter (+/- 964)`

[Update] G2: 5x speedup
- (endomorphism-based) `test bls12_381::g2::cofactor_clearing  ... bench:     264,194 ns/iter (+/- 18,145)`
- (naive, mul by cofactor) `test bls12_381::g2::cofactor_clearing  ... bench:     1,470,887 ns/iter (+/- 93,032)`
---

Before we can merge this PR, please make sure that all the following items have been
checked off. If any of the checklist items are not applicable, please leave them but
write a little note why.

- [x] Targeted PR against correct branch (master)
- [x] Linked to Github issue with discussion and accepted design OR have an explanation in the PR that describes this work.
- [x] Wrote unit tests
- [x] Updated relevant documentation in the code
- [ ] Added a relevant changelog entry to the `Pending` section in `CHANGELOG.md`
- [x] Re-reviewed `Files changed` in the Github PR explorer