@Pratyush 2022-09-06
@mmagician Could you address the CI failures and then I can re-review?

Also, it seems that the actual code multiplies by `x-1`, which in the case of BLS12-381. See, e.g., Corollary 1 [here](https://eprint.iacr.org/2021/1359.pdf)