@mmagician 2022-10-09
> Also, it seems that the actual code multiplies by `x-1`, which in the case of BLS12-381. See, e.g., Corollary 1 [here](https://eprint.iacr.org/2021/1359.pdf)

The code multiplies by 1-X, which it should, as per [WB19](https://eprint.iacr.org/2019/403.pdf). If I'm not mistaken, some later papers that are referencing the original have relied on the fact that after squaring, `1-X` and `X-1` assume the same value, so putting `c(x) = (x−1)^2 / 3 · c_2(x)` or `c(x) = (1-x)^2 / 3 · c_2(x)` is equivalent.

(Note: X is negative in curve definition.)
But the cofactor is a product of some primes, namely twice all the factors of `1-X = 15132376222941642753`, except for an extra 3:
```
sage: 15132376222941642753.factor() # effective cofactor, h_eff
3 * 11 * 10177 * 859267 * 52437899
sage: 76329603384216526031706109802092473003.factor() # cofactor, h
3 * 11^2 * 10177^2 * 859267^2 * 52437899^2
```
So actually the cofactor is `h = 15132376222941642753^2 / 3 = (1-X)^2 / 3`, and the effective cofactor should be `1-X`, again as described originally in [WB19 Section 5](https://eprint.iacr.org/2019/403.pdf).

Also this [sage script](https://gist.github.com/mmagician/567021d7609e083e250bff00317c410e) shows that indeed `X-1` doesn't work.